Implementing specification-related stuff
Basic command line options were built to accept the specification file. From that,
the JSON format validity is check, and the JSON schema that defines the specification
is applied, informing the user about the errors.

Besides, the default values are included if not present, and the types are resolved
recursively until a primitive type is reached. All over an internal representation of
the specification based on Java POJOs, automatically generated with the jsonschema2pojo
CLI.